restructuring testing stuff
