+++ src/app/features/auth/services/auth.service.spec.ts
@@ -5,7 +5,7 @@ import { Subscription } from 'rxjs/Subscription';
 
 import { AuthService } from './auth.service';
 
-import { auth } from '../../../testing/testing-helpers';
+import { auth } from '../../../testing';
 
 describe('Auth Service', () => {
   let service: AuthService;
