+++ src/app/features/tracker/components/timer/timer.component.spec.ts
@@ -19,7 +19,7 @@ import * as fromTracker from '../../reducers/root.reducer';
 
 import { TimerInfo } from '../../models';
 
-import { user } from '../../../../testing/testing-helpers';
+import { user } from '../../../../testing';
 
 describe('TimerComponent', () => {
   let store: Store<fromRoot.State>;
