+++ src/app/testing/tracker.ts
@@ -1,0 +1,5 @@
+export namespace tracker {
+  export const clockServiceStub = {
+    getCurrentTime: jasmine.createSpy('getCurrentTime')
+  };
+}
